requirements.txt
asgiref==3.6.0
bcrypt==4.0.1
Django==4.1.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
